% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter_to_bitmap.R
\name{apply_filter_to_bitmap}
\alias{apply_filter_to_bitmap}
\title{Title}
\usage{
apply_filter_to_bitmap(bitmap, mode = "greyscale", cutoff = 127)
}
\arguments{
\item{bitmap}{A 3D raw array (dimensions 3 × height × width) representing the RGB values of the image. Usually generated by \code{read_image_as_bitmap()}.}

\item{mode}{A string specifying the filter type to apply. Options: "greyscale", "red", "green", "blue", or "cutoff".}

\item{cutoff}{Integer between 0 and 255 used only when \code{mode = "cutoff"}; defaults to 127.}
}
\value{
A 3D raw array representing the filtered image bitmap.
}
\description{
Title
}
