% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuction 2.R
\name{filter_rgb}
\alias{filter_rgb}
\title{Title}
\usage{
filter_rgb(rgb_raw, mode = "greyscale", cutoff = 127)
}
\arguments{
\item{rgb_raw}{A raw vector of length 3 representing Red, Green, and Blue components.}

\item{mode}{A string indicating the filter type. Options: "greyscale", "red", "green", "blue", or "cutoff".}

\item{cutoff}{Integer between 0 and 255 used for the "cutoff" mode. Default is 127.}
}
\value{
A raw vector of 3 values with the filtered RGB result.
}
\description{
Title
}
\examples{
example
}
