% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fuction.R
\name{read_image_as_bitmap}
\alias{read_image_as_bitmap}
\title{Read image as bitmap}
\usage{
read_image_as_bitmap(image_path)
}
\arguments{
\item{image_path}{Path to the image file (e.g., JPG, PNG).}
}
\value{
A bitmap object (raw 3D RGB array).
}
\description{
Read an image file and convert it into a bitmap array that gives access to raw RGB pixel data.
}
\examples{
bmp <- read_image_as_bitmap("path/to/image.jpg")
}
